requirements.txt
py7zr==1.0.0
apache-airflow-providers-amazon==9.10.0
